training/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "blockblast-training"
version = "0.1.0"
description = "Block Blast RL training environment"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "gymnasium>=0.29",
    "pyyaml>=6.0",
    "torch>=2.0",
    "tensorboard>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "fastapi>=0.115",
    "uvicorn>=0.34",
]

[tool.setuptools]
packages = ["agent", "configs"]

[tool.pytest.ini_options]
testpaths = ["tests"]

